Fix shell usage in global completion test
Fix this test when bin/sh is a bash.  This test works fine on the CI with
the ubuntu-latest images as it uses dash for /bin/sh but on other systems, this can be a symlink to bash.

This results that on the dash, the echo command (which is an internal built
in) got executed in a way that it prints a tab character, while on bash it
prints a literal '\t' (eg a backslash and a 't'), which then makes the test
failed.

In order to fix the test, '-e' got added to echo, which interprets '\t' as
tab. It works on dash as well.